abstract cloud fetch queue stuff
